+++ app/src/main/java/com/example/playcoin/data/data_source/dto/coinDetailDTO/DeveloperData.kt
@@ -1,14 +1,15 @@
 package com.example.playcoin.data.data_source.dto.coinDetailDTO
 
+
 data class DeveloperData(
-    val closed_issues: Int,
+    val closed_issues: Double,
     val code_additions_deletions_4_weeks: CodeAdditionsDeletions4Weeks,
-    val commit_count_4_weeks: Int,
-    val forks: Int,
-    val last_4_weeks_commit_activity_series: List<Int>,
-    val pull_request_contributors: Int,
-    val pull_requests_merged: Int,
-    val stars: Int,
-    val subscribers: Int,
-    val total_issues: Int
+    val commit_count_4_weeks: Double,
+    val forks: Double,
+    val last_4_weeks_commit_activity_series: List<Double>,
+    val pull_request_contributors: Double,
+    val pull_requests_merged: Double,
+    val stars: Double,
+    val subscribers: Double,
+    val total_issues: Double
 )
